refactor: update doc comments, rename procs and files
